+++ povme/test.py
@@ -1,0 +1,13 @@
+# POVME 2.1 is released under the GNU General Public License (see
+# http://www.gnu.org/licenses/gpl.html). If you have any questions, comments,
+# or suggestions, please don't hesitate to contact me, Jacob Durrant, at
+# durrantj [at] pitt [dot] edu.
+#
+# If you use POVME in your work, please cite:
+#
+#    Durrant, J. D., C. A. de Oliveira, et al. (2011). "POVME: An algorithm
+#    for measuring binding-pocket volumes." J Mol Graph Model 29(5): 773-776.
+#
+#    Durrant, J. D., L. Votapka, et al. (2014). "Povme 2.0: An Enhanced Tool
+#    for Determining Pocket Shape and Volume Characteristics." J Chem Theory
+#    Comput 10.11: 5047-56.